Move portions of FMixerInteractivityModule that have direct dependencies on interactive-cpp to a derived type to allow for implementation of a 'null' backend.
