+++ Source/MixerInteractivity/Private/MixerInteractivityModulePrivate.cpp
@@ -12,7 +12,7 @@
 #include "MixerInteractivityUserSettings.h"
 #include "MixerDynamicDelegateBinding.h"
 #include "MixerInteractivityLog.h"
-#include "OnlineChatMixer.h"
+#include "OnlineChatMixerPrivate.h"
 
 #include "HttpModule.h"
 #include "PlatformHttp.h"
@@ -37,41 +37,13 @@
 #include "SMixerLoginPane.h"
 #endif
 
-#if PLATFORM_WINDOWS
-#include "PreWindowsApi.h"
-#define TV_API 0
-#define CPPREST_FORCE_PPLX 0
-#define XBOX_UWP 0
-#elif PLATFORM_XBOXONE
-#include "XboxOneAllowPlatformTypes.h"
-#define TV_API 1
-#endif
-#define _TURN_OFF_PLATFORM_STRING
-#define _NO_MIXERIMP
-#pragma warning(push)
-#pragma warning(disable:4628)
-#pragma warning(disable:4596)
-#pragma pack(push)
-#pragma pack(8)
-#include <interactivity_types.h>
-#include <interactivity.h>
-#pragma pack(pop)
-#pragma warning(pop)
-#if PLATFORM_WINDOWS
-#include "PostWindowsApi.h"
-#elif PLATFORM_XBOXONE
-#include "XboxOneHidePlatformTypes.h"
-#endif
-
 DEFINE_LOG_CATEGORY(LogMixerInteractivity);
-
-IMPLEMENT_MODULE(FMixerInteractivityModule, MixerInteractivity);
 
 void FMixerInteractivityModule::StartupModule()
 {
 	UserAuthState = EMixerLoginState::Not_Logged_In;
+	InteractiveConnectionAuthState = EMixerLoginState::Not_Logged_In;
 	InteractivityState = EMixerInteractivityState::Not_Interactive;
-	ClientLibraryState = Microsoft::mixer::not_initialized;
 	HasCreatedGroups = false;
 
 	ChatInterface = MakeShared<FOnlineChatMixer>();
@@ -94,7 +66,7 @@ bool FMixerInteractivityModule::LoginSilently(TSharedPtr<const FUniqueNetId> Use
 	FString Xuid = UserId->ToString();
 
 	// Go async to avoid blocking the game thread on the cross-OS call
-	PlatformUser = Async<Windows::Xbox::System::User^>(EAsyncExecution::ThreadPool,
+	XboxUserOperation = Async<Windows::Xbox::System::User^>(EAsyncExecution::ThreadPool,
 	[Xuid]() -> Windows::Xbox::System::User^
 	{
 		for (uint32 i = 0; i < Windows::Xbox::System::User::Users->Size; ++i)
@@ -114,13 +86,9 @@ bool FMixerInteractivityModule::LoginSilently(TSharedPtr<const FUniqueNetId> Use
 
 		// User is already logged in, but client library not initialized.
 		// This case will occur during PIE when logging in for interactivity with the same Mixer user that owns the Editor settings.
-		const UMixerInteractivitySettings* Settings = GetDefault<UMixerInteractivitySettings>();
-		const UMixerInteractivityUserSettings* UserSettings = GetDefault<UMixerInteractivityUserSettings>();
-		Microsoft::mixer::interactivity_manager::get_singleton_instance()->set_oauth_token(*UserSettings->AccessToken);
-		if (Microsoft::mixer::interactivity_manager::get_singleton_instance()->initialize(*FString::FromInt(Settings->GameVersionId), false, *Settings->ShareCode))
+		if (StartInteractiveConnection())
 		{
 			// Set this immediately so that polling for state matches the event
-			ClientLibraryState = Microsoft::mixer::initializing;
 			OnLoginStateChanged().Broadcast(EMixerLoginState::Logging_In);
 			return true;
 		}
@@ -319,24 +287,26 @@ EMixerLoginState FMixerInteractivityModule::GetLoginState()
 		check(CurrentUser.IsValid());
 		if (NeedsClientLibraryActive())
 		{
-			switch (ClientLibraryState)
-			{
-			case Microsoft::mixer::not_initialized:
-				return EMixerLoginState::Not_Logged_In;
+			return InteractiveConnectionAuthState;
 
-			case Microsoft::mixer::initializing:
-				return EMixerLoginState::Logging_In;
+			//switch (ClientLibraryState)
+			//{
+			//case Microsoft::mixer::not_initialized:
+			//	return EMixerLoginState::Not_Logged_In;
 
-			case Microsoft::mixer::interactivity_disabled:
-			case Microsoft::mixer::interactivity_enabled:
-			case Microsoft::mixer::interactivity_pending:
-				return EMixerLoginState::Logged_In;
+			//case Microsoft::mixer::initializing:
+			//	return EMixerLoginState::Logging_In;
 
-			default:
-				// Internal error in Mixer client library state management
-				check(false);
-				return EMixerLoginState::Not_Logged_In;
-			}
+			//case Microsoft::mixer::interactivity_disabled:
+			//case Microsoft::mixer::interactivity_enabled:
+			//case Microsoft::mixer::interactivity_pending:
+			//	return EMixerLoginState::Logged_In;
+
+			//default:
+			//	// Internal error in Mixer client library state management
+			//	check(false);
+			//	return EMixerLoginState::Not_Logged_In;
+			//}
 		}
 		else
 		{
@@ -362,34 +332,7 @@ bool FMixerInteractivityModule::Tick(float DeltaTime)
 	TickXboxLogin();
 #endif
 
-	TickParticipantCacheMaintenance();
-	TickClientLibrary();
 	TickLocalUserMaintenance();
-
-	return true;
-}
-
-void FMixerInteractivityModule::TickParticipantCacheMaintenance()
-{
-	static const FTimespan IntervalForCacheFreshness = FTimespan::FromSeconds(30.0);
-	FDateTime TimeNow = FDateTime::Now();
-	for (TMap<uint32, TSharedPtr<FMixerRemoteUserCached>>::TIterator It(RemoteParticipantCache); It; ++It)
-	{
-		FDateTime MostRecentInteraction = FMath::Max(It.Value()->ConnectedAt, It.Value()->InputAt);
-		if (!It.Value().IsUnique() || TimeNow - MostRecentInteraction < IntervalForCacheFreshness)
-		{
-			It.Value()->UpdateFromSourceParticipant();
-		}
-		else
-		{
-			It.RemoveCurrent();
-		}
-	}
-}
-
-void FMixerInteractivityModule::TickClientLibrary()
-{
-	using namespace Microsoft::mixer;
 
 	if (!NeedsClientLibraryActive())
 	{
@@ -399,144 +342,7 @@ void FMixerInteractivityModule::TickClientLibrary()
 		HasCreatedGroups = false;
 	}
 
-	std::vector<interactive_event> EventsThisFrame = interactivity_manager::get_singleton_instance()->do_work();
-	for (auto& MixerEvent : EventsThisFrame)
-	{
-		switch (MixerEvent.event_type())
-		{
-		case interactive_event_type::error:
-			// Errors that impact our login state are accompanied by an interactivity_state_changed event, so
-			// dealing with them here is just double counting.  Stick to outputting the message.
-			UE_LOG(LogMixerInteractivity, Warning, TEXT("%s"), MixerEvent.err_message().c_str());
-			break;
-
-		case interactive_event_type::interactivity_state_changed:
-		{
-			auto StateChangeArgs = std::static_pointer_cast<interactivity_state_change_event_args>(MixerEvent.event_args());
-			EMixerLoginState PreviousLoginState = GetLoginState();
-			ClientLibraryState = StateChangeArgs->new_state();
-			switch (StateChangeArgs->new_state())
-			{
-			case interactivity_state::not_initialized:
-				InteractivityState = EMixerInteractivityState::Not_Interactive;
-				switch (PreviousLoginState)
-				{
-				case EMixerLoginState::Logging_In:
-					// On Xbox a pop to not_initialized is expected as a result of calling set_local_user when there
-					// was already a previous user.  In any case on all platforms we can safely postpone dealing with
-					// the client library state until user auth is finished.
-					if (UserAuthState != EMixerLoginState::Logging_In)
-					{
-						LoginAttemptFinished(false);
-					}
-					break;
-
-				case EMixerLoginState::Logged_In:
-					// This will occur when stopping PIE.  It's annoying to have to login
-					// again to edit Mixer settings, so don't trigger logout here.
-					if (!GIsEditor)
-					{
-						Logout();
-					}
-					break;
-
-				default:
-					break;
-				}
-				break;
-
-			case interactivity_state::initializing:
-				InteractivityState = EMixerInteractivityState::Not_Interactive;
-				// Ensure the default group has a non-null representation
-				CreateGroup(NAME_DefaultMixerParticipantGroup);
-				break;
-
-			case interactivity_state::interactivity_pending:
-				if (PreviousLoginState == EMixerLoginState::Logging_In)
-				{
-					LoginAttemptFinished(true);
-				}
-				if (!HasCreatedGroups)
-				{
-					InitDesignTimeGroups();
-				}
-				break;
-
-			case interactivity_state::interactivity_disabled:
-				InteractivityState = EMixerInteractivityState::Not_Interactive;
-				if (PreviousLoginState == EMixerLoginState::Logging_In)
-				{
-					LoginAttemptFinished(true);
-				}
-				if (!HasCreatedGroups)
-				{
-					InitDesignTimeGroups();
-				}
-				break;
-
-			case interactivity_state::interactivity_enabled:
-				InteractivityState = EMixerInteractivityState::Interactive;
-				if (PreviousLoginState == EMixerLoginState::Logging_In)
-				{
-					LoginAttemptFinished(true);
-				}
-				if (!HasCreatedGroups)
-				{
-					InitDesignTimeGroups();
-				}
-				break;
-			}
-		}
-		break;
-
-		case interactive_event_type::participant_state_changed:
-		{
-			auto ParticipantEventArgs = std::static_pointer_cast<interactive_participant_state_change_event_args>(MixerEvent.event_args());
-			TSharedPtr<const FMixerRemoteUser> RemoteParticipant = CreateOrUpdateCachedParticipant(ParticipantEventArgs->participant());
-			switch (ParticipantEventArgs->state())
-			{
-			case interactive_participant_state::joined:
-				ParticipantStateChanged.Broadcast(RemoteParticipant, EMixerInteractivityParticipantState::Joined);
-				break;
-
-			case interactive_participant_state::left:
-				ParticipantStateChanged.Broadcast(RemoteParticipant, EMixerInteractivityParticipantState::Left);
-				break;
-
-			case interactive_participant_state::input_disabled:
-				ParticipantStateChanged.Broadcast(RemoteParticipant, EMixerInteractivityParticipantState::Input_Disabled);
-				break;
-
-			default:
-				break;
-			}
-		}
-		break;
-
-		case interactive_event_type::button:
-		{
-			auto OriginalButtonArgs = std::static_pointer_cast<interactive_button_event_args>(MixerEvent.event_args());
-			TSharedPtr<const FMixerRemoteUser> RemoteParticipant = CreateOrUpdateCachedParticipant(OriginalButtonArgs->participant());
-			FMixerButtonEventDetails Details;
-			Details.Pressed = OriginalButtonArgs->is_pressed();
-			Details.TransactionId = OriginalButtonArgs->transaction_id().c_str();
-			Details.SparkCost = OriginalButtonArgs->cost();
-			ButtonEvent.Broadcast(FName(OriginalButtonArgs->control_id().c_str()), RemoteParticipant, Details);
-		}
-		break;
-
-		case interactive_event_type::joystick:
-		{
-			auto OriginalStickArgs = std::static_pointer_cast<interactive_joystick_event_args>(MixerEvent.event_args());
-			TSharedPtr<const FMixerRemoteUser> RemoteParticipant = CreateOrUpdateCachedParticipant(OriginalStickArgs->participant());
-			StickEvent.Broadcast(FName(OriginalStickArgs->control_id().c_str()), RemoteParticipant, FVector2D(OriginalStickArgs->x(), OriginalStickArgs->y()));
-			break;
-		}
-
-		default:
-			break;
-		}
-	}
+	return true;
 }
 
 void FMixerInteractivityModule::TickLocalUserMaintenance()
@@ -655,7 +461,6 @@ void FMixerInteractivityModule::OnTokenRequestComplete(FHttpRequestPtr HttpReque
 				if (GotAccessToken)
 				{
 					UserSettings->SaveConfig();
-					Microsoft::mixer::interactivity_manager::get_singleton_instance()->set_oauth_token(*UserSettings->AccessToken);
 				}
 			}
 		}
@@ -702,29 +507,56 @@ void FMixerInteractivityModule::OnUserRequestComplete(FHttpRequestPtr HttpReques
 
 		if (NeedsClientLibraryActive())
 		{
-			const UMixerInteractivitySettings* Settings = GetDefault<UMixerInteractivitySettings>();
-			switch (ClientLibraryState)
+			switch (InteractiveConnectionAuthState)
 			{
-			case Microsoft::mixer::not_initialized:
-				if (!Microsoft::mixer::interactivity_manager::get_singleton_instance()->initialize(*FString::FromInt(Settings->GameVersionId), false, *Settings->ShareCode))
+			case EMixerLoginState::Not_Logged_In:
+				if (!StartInteractiveConnection())
 				{
 					LoginAttemptFinished(false);
 				}
 				else
 				{
-					// Set this immediately to avoid a temporary pop to Not_Logged_In
-					ClientLibraryState = Microsoft::mixer::initializing;
+					check(InteractiveConnectionAuthState != EMixerLoginState::Not_Logged_In);
 				}
 				break;
 
-			case Microsoft::mixer::initializing:
+			case EMixerLoginState::Logging_In:
 				// Should naturally progress and we'll handle it in Tick
 				break;
 
-			default:
+			case EMixerLoginState::Logged_In:
 				// We're already logged in
 				LoginAttemptFinished(true);
+				break;
+
+			default:
+				// Internal error in state management
+				check(false);
+				break;
 			}
+
+			//switch (ClientLibraryState)
+			//{
+			//case Microsoft::mixer::not_initialized:
+			//	if (!Microsoft::mixer::interactivity_manager::get_singleton_instance()->initialize(*FString::FromInt(Settings->GameVersionId), false, *Settings->ShareCode))
+			//	{
+			//		LoginAttemptFinished(false);
+			//	}
+			//	else
+			//	{
+			//		// Set this immediately to avoid a temporary pop to Not_Logged_In
+			//		ClientLibraryState = Microsoft::mixer::initializing;
+			//	}
+			//	break;
+
+			//case Microsoft::mixer::initializing:
+			//	// Should naturally progress and we'll handle it in Tick
+			//	break;
+
+			//default:
+			//	// We're already logged in
+			//	LoginAttemptFinished(true);
+			//}
 		}
 		else
 		{
@@ -737,101 +569,9 @@ void FMixerInteractivityModule::OnUserRequestComplete(FHttpRequestPtr HttpReques
 	}
 }
 
-void FMixerInteractivityModule::StartInteractivity()
-{
-	switch (Microsoft::mixer::interactivity_manager::get_singleton_instance()->interactivity_state())
-	{
-	case Microsoft::mixer::interactivity_disabled:
-		check(InteractivityState == EMixerInteractivityState::Not_Interactive || InteractivityState == EMixerInteractivityState::Interactivity_Stopping);
-		Microsoft::mixer::interactivity_manager::get_singleton_instance()->start_interactive();
-		InteractivityState = EMixerInteractivityState::Interactivity_Starting;
-		break;
-
-	case Microsoft::mixer::interactivity_enabled:
-	case Microsoft::mixer::interactivity_pending:
-		check(InteractivityState == EMixerInteractivityState::Interactivity_Starting || InteractivityState == EMixerInteractivityState::Interactive);
-		// No-op, but not a problem
-		break;
-
-	case Microsoft::mixer::not_initialized:
-	case Microsoft::mixer::initializing:
-		check(InteractivityState == EMixerInteractivityState::Not_Interactive || InteractivityState == EMixerInteractivityState::Interactivity_Stopping);
-		// Caller should wait!
-		// @TODO: tell them so.
-		break;
-
-	default:
-		// Internal error in state management
-		check(false);
-		break;
-	}
-}
-
-void FMixerInteractivityModule::StopInteractivity()
-{
-	switch (Microsoft::mixer::interactivity_manager::get_singleton_instance()->interactivity_state())
-	{
-	case Microsoft::mixer::interactivity_enabled:
-	case Microsoft::mixer::interactivity_pending:
-		check(InteractivityState == EMixerInteractivityState::Interactivity_Starting || 
-				InteractivityState == EMixerInteractivityState::Interactivity_Stopping || 
-				InteractivityState == EMixerInteractivityState::Interactive);
-		Microsoft::mixer::interactivity_manager::get_singleton_instance()->stop_interactive();
-		InteractivityState = EMixerInteractivityState::Interactivity_Stopping;
-		break;
-
-	case Microsoft::mixer::interactivity_disabled:
-		check(InteractivityState == EMixerInteractivityState::Not_Interactive || InteractivityState == EMixerInteractivityState::Interactivity_Stopping);
-		// No-op, but not a problem
-		break;
-
-	case Microsoft::mixer::not_initialized:
-	case Microsoft::mixer::initializing:
-		check(InteractivityState == EMixerInteractivityState::Not_Interactive || InteractivityState == EMixerInteractivityState::Interactivity_Stopping);
-		// Caller should wait!
-		// @TODO: tell them so.
-		break;
-
-	default:
-		// Internal error in state management
-		check(false);
-		break;
-	}
-}
-
 EMixerInteractivityState FMixerInteractivityModule::GetInteractivityState()
 {
 	return InteractivityState;
-}
-
-void FMixerInteractivityModule::SetCurrentScene(FName Scene, FName GroupName)
-{
-	using namespace Microsoft::mixer;
-
-	if (GetInteractivityState() == EMixerInteractivityState::Interactive)
-	{
-		std::shared_ptr<interactive_group> Group = GroupName == NAME_None ? interactivity_manager::get_singleton_instance()->group() : interactivity_manager::get_singleton_instance()->group(*GroupName.ToString());
-		std::shared_ptr<interactive_scene> TargetScene = interactivity_manager::get_singleton_instance()->scene(*Scene.ToString());
-		if (Group != nullptr && TargetScene != nullptr)
-		{
-			Group->set_scene(TargetScene);
-		}
-	}
-}
-
-FName FMixerInteractivityModule::GetCurrentScene(FName GroupName)
-{
-	using namespace Microsoft::mixer;
-	FName SceneName = NAME_None;
-	if (GetInteractivityState() == EMixerInteractivityState::Interactive)
-	{
-		std::shared_ptr<interactive_group> Group = GroupName == NAME_None ? interactivity_manager::get_singleton_instance()->group() : interactivity_manager::get_singleton_instance()->group(*GroupName.ToString());
-		if (Group && Group->scene())
-		{
-			SceneName = Group->scene()->scene_id().c_str();
-		}
-	}
-	return SceneName;
 }
 
 bool FMixerInteractivityModule::NeedsClientLibraryActive()
@@ -853,192 +593,6 @@ bool FMixerInteractivityModule::NeedsClientLibraryActive()
 #else
 	return true;
 #endif
-}
-
-void FMixerInteractivityModule::TriggerButtonCooldown(FName Button, FTimespan CooldownTime)
-{
-	using namespace Microsoft::mixer;
-
-	if (GetInteractivityState() == EMixerInteractivityState::Interactive)
-	{
-		std::chrono::milliseconds CooldownTimeInMs = std::chrono::duration_cast<std::chrono::milliseconds>(std::chrono::duration<double, std::milli>(CooldownTime.GetTotalMilliseconds()));
-		interactivity_manager::get_singleton_instance()->get_singleton_instance()->trigger_cooldown(*Button.ToString(), CooldownTimeInMs);
-	}
-}
-
-bool FMixerInteractivityModule::GetButtonDescription(FName Button, FMixerButtonDescription& OutDesc)
-{
-	using namespace Microsoft::mixer;
-
-	std::shared_ptr<interactive_button_control> ButtonControl = FindButton(Button);
-	if (ButtonControl)
-	{
-		OutDesc.ButtonText = FText::FromString(ButtonControl->button_text().c_str());
-		OutDesc.HelpText = FText::GetEmpty(); //FText::FromString(ButtonControl->help_text().c_str());
-		OutDesc.SparkCost = ButtonControl->cost();
-		return true;
-	}
-	return false;
-}
-
-bool FMixerInteractivityModule::GetButtonState(FName Button, FMixerButtonState& OutState)
-{
-	using namespace Microsoft::mixer;
-
-	std::shared_ptr<interactive_button_control> ButtonControl = FindButton(Button);
-	if (ButtonControl)
-	{
-		OutState.RemainingCooldown = FTimespan::FromMilliseconds(ButtonControl->remaining_cooldown().count());
-		OutState.Progress = ButtonControl->progress();
-		OutState.PressCount = ButtonControl->count_of_button_presses();
-		OutState.DownCount = ButtonControl->count_of_button_downs();
-		OutState.UpCount = ButtonControl->count_of_button_ups();
-		OutState.Enabled = !ButtonControl->disabled();
-		return true;
-	}
-	return false;
-}
-
-bool FMixerInteractivityModule::GetButtonState(FName Button, uint32 ParticipantId, FMixerButtonState& OutState)
-{
-	using namespace Microsoft::mixer;
-
-	std::shared_ptr<interactive_button_control> ButtonControl = FindButton(Button);
-	if (ButtonControl)
-	{
-		OutState.RemainingCooldown = FTimespan::FromMilliseconds(ButtonControl->remaining_cooldown().count());
-		OutState.Progress = ButtonControl->progress();
-		OutState.PressCount = ButtonControl->is_pressed(ParticipantId) ? 1 : 0;
-		OutState.DownCount = ButtonControl->is_down(ParticipantId) ? 1 : 0;
-		OutState.UpCount = ButtonControl->is_up(ParticipantId) ? 1 : 0;
-		OutState.Enabled = !ButtonControl->disabled();
-		return true;
-	}
-	return false;
-}
-
-bool FMixerInteractivityModule::GetStickDescription(FName Stick, FMixerStickDescription& OutDesc)
-{
-	using namespace Microsoft::mixer;
-
-	std::shared_ptr<interactive_joystick_control> StickControl = FindStick(Stick);
-	if (StickControl)
-	{
-		OutDesc.HelpText = FText::GetEmpty(); //FText::FromString(StickControl->help_text().c_str());
-		return true;
-	}
-	return false;
-}
-
-bool FMixerInteractivityModule::GetStickState(FName Stick, FMixerStickState& OutState)
-{
-	using namespace Microsoft::mixer;
-
-	std::shared_ptr<interactive_joystick_control> StickControl = FindStick(Stick);
-	if (StickControl)
-	{
-		OutState.Axes = FVector2D(static_cast<float>(StickControl->x()), static_cast<float>(StickControl->y()));
-		OutState.Enabled = true; //!StickControl->disabled();
-		return true;
-	}
-	return false;
-}
-
-bool FMixerInteractivityModule::GetStickState(FName Stick, uint32 ParticipantId, FMixerStickState& OutState)
-{
-	using namespace Microsoft::mixer;
-
-	std::shared_ptr<interactive_joystick_control> StickControl = FindStick(Stick);
-	if (StickControl)
-	{
-		OutState.Axes = FVector2D(static_cast<float>(StickControl->x(ParticipantId)), static_cast<float>(StickControl->y(ParticipantId)));
-		OutState.Enabled = true; //!StickControl->disabled();
-		return true;
-	}
-	return false;
-}
-
-std::shared_ptr<Microsoft::mixer::interactive_button_control> FMixerInteractivityModule::FindButton(FName Name)
-{
-	using namespace Microsoft::mixer;
-
-	if (GetInteractivityState() == EMixerInteractivityState::Interactive)
-	{
-		FString NameAsString = Name.ToString();
-		for (const std::shared_ptr<interactive_scene> SceneObject : interactivity_manager::get_singleton_instance()->scenes())
-		{
-			if (SceneObject)
-			{
-				std::shared_ptr<Microsoft::mixer::interactive_button_control> ButtonControl = SceneObject->button(*NameAsString);
-				if (ButtonControl)
-				{
-					return ButtonControl;
-				}
-			}
-		}
-	}
-	
-	return nullptr;
-}
-
-std::shared_ptr<Microsoft::mixer::interactive_joystick_control> FMixerInteractivityModule::FindStick(FName Name)
-{
-	using namespace Microsoft::mixer;
-
-	if (GetInteractivityState() == EMixerInteractivityState::Interactive)
-	{
-		FString NameAsString = Name.ToString();
-		for (const std::shared_ptr<interactive_scene> SceneObject : interactivity_manager::get_singleton_instance()->scenes())
-		{
-			if (SceneObject)
-			{
-				std::shared_ptr<Microsoft::mixer::interactive_joystick_control> StickControl = SceneObject->joystick(*NameAsString);
-				if (StickControl)
-				{
-					return StickControl;
-				}
-			}
-		}
-	}
-
-	return nullptr;
-}
-
-TSharedPtr<const FMixerRemoteUser> FMixerInteractivityModule::GetParticipant(uint32 ParticipantId)
-{
-	using namespace Microsoft::mixer;
-
-	if (GetInteractivityState() == EMixerInteractivityState::Interactive)
-	{
-		TSharedPtr<FMixerRemoteUserCached>* CachedUser = RemoteParticipantCache.Find(ParticipantId);
-		if (CachedUser)
-		{
-			return *CachedUser;
-		}
-
-		for (std::shared_ptr<interactive_participant> Participant : interactivity_manager::get_singleton_instance()->participants())
-		{
-			check(Participant);
-			if (Participant->mixer_id() == ParticipantId)
-			{
-				return CreateOrUpdateCachedParticipant(Participant);
-			}
-		}
-	}
-
-	return nullptr;
-}
-
-TSharedPtr<FMixerRemoteUserCached> FMixerInteractivityModule::CreateOrUpdateCachedParticipant(std::shared_ptr<Microsoft::mixer::interactive_participant> Participant)
-{
-	check(Participant);
-	TSharedPtr<FMixerRemoteUserCached>& NewUser = RemoteParticipantCache.Add(Participant->mixer_id());
-	if (!NewUser.IsValid())
-	{
-		NewUser = MakeShareable(new FMixerRemoteUserCached(Participant));
-	}
-	NewUser->UpdateFromSourceParticipant();
-	return NewUser;
 }
 
 void FMixerInteractivityModule::LoginAttemptFinished(bool Success)
@@ -1086,103 +640,7 @@ void FMixerInteractivityModule::LoginAttemptFinished(bool Success)
 	}
 }
 
-bool FMixerInteractivityModule::CreateGroup(FName GroupName, FName InitialScene)
-{
-	using namespace Microsoft::mixer;
 
-	FString GroupNameAsString = GroupName.ToString();
-	std::shared_ptr<interactive_group> FoundGroup = interactivity_manager::get_singleton_instance()->group(*GroupName.ToString());
-	bool CanCreate = FoundGroup == nullptr;
-	if (CanCreate)
-	{
-		if (InitialScene != NAME_None)
-		{
-			std::shared_ptr<interactive_scene> TargetScene = interactivity_manager::get_singleton_instance()->scene(*InitialScene.ToString());
-			if (TargetScene)
-			{
-				std::make_shared<interactive_group>(*GroupNameAsString, TargetScene);
-			}
-			else
-			{
-				CanCreate = false;
-			}
-		}
-		else
-		{
-			// Constructor adds it to the internal manager.
-			std::make_shared<interactive_group>(*GroupNameAsString);
-		}
-	}
-
-	return CanCreate;
-}
-
-bool FMixerInteractivityModule::GetParticipantsInGroup(FName GroupName, TArray<TSharedPtr<const FMixerRemoteUser>>& OutParticipants)
-{
-	using namespace Microsoft::mixer;
-
-	std::shared_ptr<interactive_group> ExistingGroup = interactivity_manager::get_singleton_instance()->group(*GroupName.ToString());
-	bool FoundGroup = false;
-	if (ExistingGroup)
-	{
-		FoundGroup = true;
-		const std::vector<std::shared_ptr<interactive_participant>> ParticipantsInternal = ExistingGroup->participants();
-		OutParticipants.Empty(ParticipantsInternal.size());
-		for (std::shared_ptr<interactive_participant> Participant : ParticipantsInternal)
-		{
-			OutParticipants.Add(CreateOrUpdateCachedParticipant(Participant));
-		}
-	}
-	else
-	{
-		OutParticipants.Empty();
-	}
-
-	return FoundGroup;
-}
-
-bool FMixerInteractivityModule::MoveParticipantToGroup(FName GroupName, uint32 ParticipantId)
-{
-	using namespace Microsoft::mixer;
-
-	FString GroupNameAsString = GroupName.ToString();
-	std::shared_ptr<interactive_group> ExistingGroup = interactivity_manager::get_singleton_instance()->group(*GroupNameAsString);
-	bool FoundUser = false;
-	if (ExistingGroup)
-	{
-		std::shared_ptr<interactive_participant> Participant;
-		TSharedPtr<FMixerRemoteUserCached>* CachedUser = RemoteParticipantCache.Find(ParticipantId);
-		if (CachedUser)
-		{
-			Participant = (*CachedUser)->GetSourceParticipant();
-		}
-		else
-		{
-			for (std::shared_ptr<interactive_participant> PossibleParticipant : interactivity_manager::get_singleton_instance()->participants())
-			{
-				check(PossibleParticipant);
-				if (PossibleParticipant->mixer_id() == ParticipantId)
-				{
-					Participant = PossibleParticipant;
-					break;
-				}
-			}
-		}
-
-		if (Participant)
-		{
-			FoundUser = true;
-			Participant->set_group(ExistingGroup);
-			CreateOrUpdateCachedParticipant(Participant);
-		}
-	}
-	return FoundUser;
-}
-
-void FMixerInteractivityModule::CaptureSparkTransaction(const FString& TransactionId)
-{
-	Microsoft::mixer::interactivity_manager::get_singleton_instance()->capture_transaction(*TransactionId);
-}
 
 void FMixerInteractivityModule::InitDesignTimeGroups()
 {
@@ -1243,13 +701,11 @@ void FMixerInteractivityModule::TickXboxLogin()
 				GetXTokenOperation = nullptr;
 			}
 		}
-		else  if (PlatformUser.IsReady())
+		else  if (XboxUserOperation.IsReady())
 		{
-			Windows::Xbox::System::User^ ResolvedUser = PlatformUser.Get();
+			Windows::Xbox::System::User^ ResolvedUser = GetXboxUser();
 			if (ResolvedUser != nullptr)
 			{
-				Microsoft::mixer::interactivity_manager::get_singleton_instance()->set_local_user(PlatformUser.Get());
-
 				try
 				{
 					GetXTokenOperation = ResolvedUser->GetTokenAndSignatureAsync(L"POST", L"https://mixer.com", L"");
@@ -1263,8 +719,6 @@ void FMixerInteractivityModule::TickXboxLogin()
 			{
 				LoginError = true;
 			}
-
-			PlatformUser = TFuture<Windows::Xbox::System::User^>();
 		}
 
 		if (LoginError)
@@ -1277,23 +731,3 @@ void FMixerInteractivityModule::TickXboxLogin()
 	}
 }
 #endif
-
-FMixerRemoteUserCached::FMixerRemoteUserCached(std::shared_ptr<Microsoft::mixer::interactive_participant> InParticipant)
-	: SourceParticipant(InParticipant)
-{
-	Id = SourceParticipant->mixer_id();
-}
-
-void FMixerRemoteUserCached::UpdateFromSourceParticipant()
-{
-	// Is there really not a std definition for this?
-	typedef std::chrono::duration<uint64, std::ratio_multiply<std::nano, std::ratio<100>>> DateTimeTicks;
-
-	Name = SourceParticipant->username().c_str();
-	Level = SourceParticipant->level();
-	ConnectedAt = FDateTime::FromUnixTimestamp(std::chrono::duration_cast<DateTimeTicks>(SourceParticipant->connected_at()).count());
-	InputAt = FDateTime::FromUnixTimestamp(std::chrono::duration_cast<DateTimeTicks>(SourceParticipant->last_input_at()).count());
-	InputEnabled = !SourceParticipant->input_disabled();
-	std::shared_ptr<Microsoft::mixer::interactive_group> GroupInternal = SourceParticipant->group();
-	Group = GroupInternal ? FName(GroupInternal->group_id().c_str()) : NAME_DefaultMixerParticipantGroup;
-}
