+++ Source/MixerInteractivity/Private/MixerInteractivityModulePrivate.h
@@ -30,17 +30,6 @@
 #include "CoreOnline.h"
 #include <memory>
 
-namespace Microsoft
-{
-	namespace mixer
-	{
-		class interactive_button_control;
-		class interactive_joystick_control;
-		class interactive_participant;
-		enum interactivity_state : int;
-	}
-}
-
 struct FMixerChannelJsonSerializable : public FMixerChannel, public FJsonSerializable
 {
 public:
@@ -77,18 +66,6 @@ public:
 	}
 };
 
-struct FMixerRemoteUserCached : public FMixerRemoteUser
-{
-public:
-	FMixerRemoteUserCached(std::shared_ptr<Microsoft::mixer::interactive_participant> InParticipant);
-
-	void UpdateFromSourceParticipant();
-
-	std::shared_ptr<Microsoft::mixer::interactive_participant> GetSourceParticipant() { return SourceParticipant; }
-private:
-	std::shared_ptr<Microsoft::mixer::interactive_participant> SourceParticipant;
-};
-
 class SWindow;
 class SOverlay;
 class IWebBrowserWindow;
@@ -109,30 +86,12 @@ public:
 	virtual bool Logout();
 	virtual EMixerLoginState GetLoginState();
 
-	virtual void StartInteractivity();
-	virtual void StopInteractivity();
 	virtual EMixerInteractivityState GetInteractivityState();
-
-	virtual void SetCurrentScene(FName Scene, FName GroupName = NAME_None);
-	virtual FName GetCurrentScene(FName GroupName = NAME_None);
-	virtual void TriggerButtonCooldown(FName Button, FTimespan CooldownTime);
-	virtual bool GetButtonDescription(FName Button, FMixerButtonDescription& OutDesc);
-	virtual bool GetButtonState(FName Button, FMixerButtonState& OutState);
-	virtual bool GetButtonState(FName Button, uint32 ParticipantId, FMixerButtonState& OutState);
-	virtual bool GetStickDescription(FName Stick, FMixerStickDescription& OutDesc);
-	virtual bool GetStickState(FName Stick, FMixerStickState& OutState);
-	virtual bool GetStickState(FName Stick, uint32 ParticipantId, FMixerStickState& OutState);
 
 	virtual TSharedPtr<const FMixerLocalUser> GetCurrentUser()
 	{
 		return CurrentUser;
 	}
-	virtual TSharedPtr<const FMixerRemoteUser> GetParticipant(uint32 ParticipantId);
-
-	virtual bool CreateGroup(FName GroupName, FName InitialScene = NAME_None);
-	virtual bool GetParticipantsInGroup(FName GroupName, TArray<TSharedPtr<const FMixerRemoteUser>>& OutParticipants);
-	virtual bool MoveParticipantToGroup(FName GroupName, uint32 ParticipantId);
-	virtual void CaptureSparkTransaction(const FString& TransactionId);
 
 	virtual TSharedPtr<class IOnlineChat> GetChatInterface();
 	virtual TSharedPtr<class IOnlineChatMixer> GetExtendedChatInterface();
@@ -145,8 +104,23 @@ public:
 	virtual FOnBroadcastingStateChanged& OnBroadcastingStateChanged()		{ return BroadcastingStateChanged; }
 
 public:
-
 	virtual bool Tick(float DeltaTime);
+
+protected:
+	virtual bool StartInteractiveConnection() = 0;
+
+protected:
+	void LoginAttemptFinished(bool Success);
+
+	EMixerLoginState GetUserAuthState() const							{ return UserAuthState; }
+	EMixerLoginState GetInteractiveConntectionAuthState() const			{ return InteractiveConnectionAuthState; }
+	void SetInteractiveConnectionAuthState(EMixerLoginState InState)	{ InteractiveConnectionAuthState = InState; }
+	EMixerInteractivityState GetInteractivityState() const				{ return InteractivityState; }
+	void SetInteractivityState(EMixerInteractivityState InState)		{ InteractivityState = InState; InteractivityStateChanged.Broadcast(InState); }
+
+#if PLATFORM_XBOXONE
+	Windows::Xbox::System::User^ GetXboxUser()							{ return XboxUserOperation.Get(); }
+#endif
 
 private:
 	bool LoginWithAuthCodeInternal(const FString& AuthCode, TSharedPtr<const FUniqueNetId> UserId);
@@ -162,20 +136,13 @@ private:
 	bool NeedsClientLibraryActive();
 	void InitDesignTimeGroups();
 
-	std::shared_ptr<Microsoft::mixer::interactive_button_control> FindButton(FName Name);
-	std::shared_ptr<Microsoft::mixer::interactive_joystick_control> FindStick(FName Name);
-	TSharedPtr<FMixerRemoteUserCached> CreateOrUpdateCachedParticipant(std::shared_ptr<Microsoft::mixer::interactive_participant> Participant);
-
-	void TickParticipantCacheMaintenance();
-	void TickClientLibrary();
 	void TickLocalUserMaintenance();
 
-	void LoginAttemptFinished(bool Success);
 
 private:
 
 #if PLATFORM_XBOXONE
-	TFuture<Windows::Xbox::System::User^> PlatformUser;
+	TFuture<Windows::Xbox::System::User^> XboxUserOperation;
 	Windows::Foundation::IAsyncOperation<Windows::Xbox::System::GetTokenAndSignatureResult^>^ GetXTokenOperation;
 	void TickXboxLogin();
 #endif
@@ -186,10 +153,8 @@ private:
 	TSharedPtr<const FUniqueNetId> NetId;
 	TSharedPtr<FMixerLocalUserJsonSerializable> CurrentUser;
 
-	TMap<uint32, TSharedPtr<FMixerRemoteUserCached>> RemoteParticipantCache;
-
 	EMixerLoginState UserAuthState;
-	Microsoft::mixer::interactivity_state ClientLibraryState;
+	EMixerLoginState InteractiveConnectionAuthState;
 	EMixerInteractivityState InteractivityState;
 
 	FOnLoginStateChanged LoginStateChanged;
