+++ Source/MixerInteractivity/Private/MixerInteractivityModule_InteractiveCpp.h
@@ -1,0 +1,66 @@
+#pragma once
+
+#include "MixerInteractivityModulePrivate.h"
+
+namespace Microsoft
+{
+	namespace mixer
+	{
+		class interactive_button_control;
+		class interactive_joystick_control;
+		class interactive_participant;
+		enum interactivity_state : int;
+	}
+}
+
+struct FMixerRemoteUserCached : public FMixerRemoteUser
+{
+public:
+	FMixerRemoteUserCached(std::shared_ptr<Microsoft::mixer::interactive_participant> InParticipant);
+
+	void UpdateFromSourceParticipant();
+
+	std::shared_ptr<Microsoft::mixer::interactive_participant> GetSourceParticipant() { return SourceParticipant; }
+private:
+	std::shared_ptr<Microsoft::mixer::interactive_participant> SourceParticipant;
+};
+
+
+class FMixerInteractivityModule_InteractiveCpp : public FMixerInteractivityModule
+{
+public:
+	virtual void StartInteractivity();
+	virtual void StopInteractivity();
+	virtual void SetCurrentScene(FName Scene, FName GroupName = NAME_None);
+	virtual FName GetCurrentScene(FName GroupName = NAME_None);
+	virtual void TriggerButtonCooldown(FName Button, FTimespan CooldownTime);
+	virtual bool GetButtonDescription(FName Button, FMixerButtonDescription& OutDesc);
+	virtual bool GetButtonState(FName Button, FMixerButtonState& OutState);
+	virtual bool GetButtonState(FName Button, uint32 ParticipantId, FMixerButtonState& OutState);
+	virtual bool GetStickDescription(FName Stick, FMixerStickDescription& OutDesc);
+	virtual bool GetStickState(FName Stick, FMixerStickState& OutState);
+	virtual bool GetStickState(FName Stick, uint32 ParticipantId, FMixerStickState& OutState);
+	virtual TSharedPtr<const FMixerRemoteUser> GetParticipant(uint32 ParticipantId);
+	virtual bool CreateGroup(FName GroupName, FName InitialScene = NAME_None);
+	virtual bool GetParticipantsInGroup(FName GroupName, TArray<TSharedPtr<const FMixerRemoteUser>>& OutParticipants);
+	virtual bool MoveParticipantToGroup(FName GroupName, uint32 ParticipantId);
+	virtual void CaptureSparkTransaction(const FString& TransactionId);
+
+public:
+	virtual bool Tick(float DeltaTime) override;
+
+protected:
+	virtual bool StartInteractiveConnection();
+
+private:
+	std::shared_ptr<Microsoft::mixer::interactive_button_control> FindButton(FName Name);
+	std::shared_ptr<Microsoft::mixer::interactive_joystick_control> FindStick(FName Name);
+	TSharedPtr<FMixerRemoteUserCached> CreateOrUpdateCachedParticipant(std::shared_ptr<Microsoft::mixer::interactive_participant> Participant);
+
+	void TickParticipantCacheMaintenance();
+
+private:
+	TMap<uint32, TSharedPtr<FMixerRemoteUserCached>> RemoteParticipantCache;
+	Microsoft::mixer::interactivity_state ClientLibraryState;
+};
+
